[STYLE/HM-94]일반 컨테이너 스타일 추가 - 중앙으로 요소를 모으지 않는 버전 / 결과페이지 일정조율 페이지에 일반 컨테이너추가 / 바텀시트 컴포넌트 분리 및 디자인 조
