+++ src/components/meeting/result/AttendStatusHeader.tsx
@@ -1,11 +1,7 @@
-import Button from '@/components/common/Button';
-import {
-  BottomSheetHeader,
-  BottomSheetTitle,
-} from '@/styles/components/bottomsheet/bottomsheet';
-import theme from '@/styles/theme';
 import { useState } from 'react';
-import styled, { keyframes } from 'styled-components';
+import styled from 'styled-components';
+import theme from '@/styles/theme';
+import ResultBottomSheet from './ResultBottomSheet';
 
 function AttendStatusHeader({
   TotalParticipants,
@@ -20,11 +16,8 @@ function AttendStatusHeader({
 }) {
   const [isBottomSheetOpen, setIsBottomSheetOpen] = useState(false);
 
-  const participatedUsersList = participatedUsers.join(', ');
-  const unParticipatedUsersList = unParticipatedUsers.join(', ');
-  const toggleBottomSheet = () => {
-    setIsBottomSheetOpen(!isBottomSheetOpen);
-  };
+  const openBottomSheet = () => setIsBottomSheetOpen(true);
+  const closeBottomSheet = () => setIsBottomSheetOpen(false);
 
   return (
     <>
@@ -37,54 +30,24 @@ function AttendStatusHeader({
           <Title>일정 조율 중</Title>
           <AttendStatus>
             (현재 참여인원 : {TotalParticipants}명 중) 참여자
-            <AttendParticipantCount onClick={toggleBottomSheet}>
+            <AttendParticipantCount onClick={openBottomSheet}>
               {currentParticipants}명
             </AttendParticipantCount>
           </AttendStatus>
         </AttendStatusHeaderContainer>
       )}
       {isBottomSheetOpen && (
-        <BottomSheetContainer>
-          <BottomSheetHeader>
-            <BottomSheetTitle>현재 참여 인원</BottomSheetTitle>
-            <CloseButton onClick={toggleBottomSheet}>X</CloseButton>
-          </BottomSheetHeader>
-          <UserCountContainer>
-            <UserState>
-              총 참여 인원 :
-              <ParticipantCount>{TotalParticipants}명</ParticipantCount>
-            </UserState>
-            <VerticalDivider />
-            <UserState>참여 완료 : {currentParticipants}</UserState>
-            <UserState>
-              미참여 : {TotalParticipants - currentParticipants}
-            </UserState>
-          </UserCountContainer>
-          <UserListContainer>
-            <ParticipantIcon participant>참여 완료</ParticipantIcon>
-            <ParticipantList participant>
-              {participatedUsersList}
-            </ParticipantList>
-          </UserListContainer>
-          <UserListContainer>
-            <ParticipantIcon participant={false}>미참여</ParticipantIcon>
-            <ParticipantList participant={false}>
-              {unParticipatedUsersList}
-            </ParticipantList>
-          </UserListContainer>
-          <Button $style="solid">일정 조율 완료</Button>
-        </BottomSheetContainer>
+        <ResultBottomSheet
+          TotalParticipants={TotalParticipants}
+          currentParticipants={currentParticipants}
+          participatedUsers={participatedUsers}
+          unParticipatedUsers={unParticipatedUsers}
+          onClose={closeBottomSheet}
+        />
       )}
     </>
   );
 }
-
-const colors = {
-  participantText: '#4c545c',
-  participantBackground: 'rgba(226, 245, 227, 1)',
-  nonParticipantText: 'rgba(255, 114, 83, 0.3)',
-  nonParticipantBackground: 'rgba(255, 114, 83, 0.3)',
-};
 
 const AttendStatusHeaderContainer = styled.div`
   display: flex;
@@ -112,95 +75,4 @@ const AttendParticipantCount = styled.span`
   cursor: pointer;
 `;
 
-const CloseButton = styled.button`
-  font-size: 16px;
-  font-weight: bold;
-  color: #33c894;
-  background: none;
-  border: none;
-  cursor: pointer;
-`;
-const UserCountContainer = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  padding: 15px 0;
-`;
-
-const VerticalDivider = styled.div`
-  width: 1px;
-  height: 40px;
-  background-color: #e0e0e0;
-  margin: 0 15px;
-`;
-
-const UserState = styled.div`
-  font-size: 19px;
-  display: flex;
-  align-items: center;
-`;
-
-const ParticipantCount = styled.span`
-  font-size: 18px;
-  font-weight: bold;
-  margin-left: 10px;
-`;
-const UserListContainer = styled.div`
-  display: flex;
-  width: 100%;
-  gap: 10px;
-  flex-direction: column;
-  margin: 15px;
-`;
-
-const ParticipantList = styled.div<{ participant: boolean }>`
-  font-size: 14px;
-  font-weight: bold;
-  color: ${(props) =>
-    props.participant ? colors.participantText : colors.nonParticipantText};
-`;
-
-const ParticipantIcon = styled.div<{ participant: boolean }>`
-  background-color: ${(props) =>
-    props.participant
-      ? colors.participantBackground
-      : colors.nonParticipantBackground};
-  border-radius: 40px;
-  padding: 5px 10px;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  color: ${colors.participantText};
-  font-weight: bold;
-  font-size: 13px;
-  min-width: 30px;
-  max-width: 100px;
-`;
-
-const slideIn = keyframes`
-  0% { transform: translateY(100%); }
-  100% { transform: translateY(0); }
-`;
-
-export const BottomSheetContainer = styled.div`
-  width: 100%;
-  height: 50vh;
-
-  position: absolute;
-  bottom: 0;
-
-  margin-top: 22%;
-  padding: 24px;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 5px;
-
-  border-radius: 20px 20px 0px 0px;
-  background: ${theme.color.primary.white};
-
-  animation: ${slideIn} 240ms cubic-bezier(0.5, 0.1, 0.34, 1);
-`;
 export default AttendStatusHeader;
